+++ Common/src/main/java/com/almostreliable/unified/compat/RecipeIndicator.java
@@ -15,12 +15,13 @@ import java.util.List;
 
 public final class RecipeIndicator {
 
-    public static final int SIZE = 16;
+    public static final int RENDER_SIZE = 10;
+    private static final int TEXTURE_SIZE = 16;
     private static final ResourceLocation TEXTURE = Utils.getRL("textures/ingot.png");
 
     private RecipeIndicator() {}
 
-    public static List<Component> constructTooltip(ClientRecipeLink link) {
+    static List<Component> constructTooltip(ClientRecipeLink link) {
         var unified = Component.translatable(Utils.prefix("unified")).append(": ")
                 .withStyle(c -> c.withColor(ChatFormatting.AQUA));
         unified.append(Component.translatable(Utils.prefix(link.isUnified() ? "yes" : "no"))
@@ -41,14 +42,14 @@ public final class RecipeIndicator {
         );
     }
 
-    public static void renderIndicator(PoseStack poseStack, Rect2i area) {
+    static void renderIndicator(PoseStack poseStack, Rect2i area) {
         poseStack.pushPose();
         poseStack.translate(area.getX(), area.getY(), 0);
-        var scale = area.getWidth() / (float) SIZE;
+        var scale = area.getWidth() / (float) TEXTURE_SIZE;
         poseStack.scale(scale, scale, scale);
         RenderSystem.setShader(GameRenderer::getPositionTexShader);
         RenderSystem.setShaderTexture(0, TEXTURE);
-        GuiComponent.blit(poseStack, 0, 0, 0, 0, SIZE, SIZE, SIZE, SIZE);
+        GuiComponent.blit(poseStack, 0, 0, 0, 0, TEXTURE_SIZE, TEXTURE_SIZE, TEXTURE_SIZE, TEXTURE_SIZE);
         poseStack.popPose();
     }
 }